currencyconverter
Simple currency converter, supports 52 different currencies. Run by web scraping the current currency USD conversion values from a live stock site. Asks the user which currency they would like to convert from, and checks it against an "available currency" list. Does the same with the converting to currency. Checks if the amount of the from currency is an integer. Then, the conversion- works by converting the from currency to its value in USD, then converting the USD value into its value in the to currency.